+++ packages/xchain-thorchain-amm/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bump Dep version number for `xchain-midgard`, `thorchain-query`
 - Change ExecuteSwap() parameters to use constructed memo from `estimateSwap()`
 - Remove `constructMemo()` from wallet.ts
+- validate affiliate address is either a valid thorchain adress OR a valid thorname
 - Clean types file
 
 # v0.3.3 (2022-10-17)